Update script.js
Adds a function (addToList) for the button press, allowing for new list elements to be added to the unordered list. 
Also, turns the addEventListener loop from the previous patch and the getElements process into functions so that they can be called repeatedly allowing for the javascript to add EventListeners to any new list elements and their children.